+++ pkg/monitoring/metrics_test.go
@@ -73,6 +73,20 @@ func TestMetricAdder(t *testing.T) {
 	require.NoError(err)
 	require.Equal("ready", status)
 
+	now := metav1.Now()
+	pods[1].SetDeletionTimestamp(&now)
+	UpdatePodStatusMetric(pods[1])
+	status, err = getCurrentPodStatus("pod1")
+	require.NoError(err)
+	require.Equal("terminating", status)
+
+	// Decommissioning should be preferred to Terminating if we are decommissioning the pod
+	pods[5].SetDeletionTimestamp(&now)
+	UpdatePodStatusMetric(pods[1])
+	status, err = getCurrentPodStatus("pod5")
+	require.NoError(err)
+	require.Equal("decommissioning", status)
+
 	RemoveDatacenterPods("ns", "cluster1", "datacenter1")
 	_, err = getCurrentPodStatus("pod4")
 	require.Error(err)
@@ -165,13 +179,13 @@ func TestOperatorStateMetrics(t *testing.T) {
 		Status: api.CassandraDatacenterStatus{},
 	}
 
-	UpdateOperatorDatacenterProgressStatusMetric(dc, api.ProgressUpdating)
+	updateOperatorDatacenterProgressStatusMetric(dc, api.ProgressUpdating)
 
 	status, err := getCurrentDatacenterStatus("dc1")
 	require.NoError(err)
 	require.Equal("Updating", status)
 
-	UpdateOperatorDatacenterProgressStatusMetric(dc, api.ProgressReady)
+	updateOperatorDatacenterProgressStatusMetric(dc, api.ProgressReady)
 
 	status, err = getCurrentDatacenterStatus("dc1")
 	require.NoError(err)
@@ -229,14 +243,14 @@ func TestDatacenterConditionMetrics(t *testing.T) {
 		},
 	}
 
-	SetDatacenterConditionMetric(dc, api.DatacenterReady, corev1.ConditionTrue)
+	setDatacenterConditionMetric(dc, api.DatacenterReady, corev1.ConditionTrue)
 
 	status, err := getCurrentDatacenterCondition("dc1", api.DatacenterReady)
 	require.NoError(err)
 	require.Equal(float64(1), status)
 
-	SetDatacenterConditionMetric(dc, api.DatacenterInitialized, corev1.ConditionTrue)
-	SetDatacenterConditionMetric(dc, api.DatacenterReady, corev1.ConditionFalse)
+	setDatacenterConditionMetric(dc, api.DatacenterInitialized, corev1.ConditionTrue)
+	setDatacenterConditionMetric(dc, api.DatacenterReady, corev1.ConditionFalse)
 
 	status, err = getCurrentDatacenterCondition("dc1", api.DatacenterReady)
 	require.NoError(err)